requirements.txt
chromadb==0.5.5
google-generativeai==0.7.2
langchain==0.2.13
langchain-chroma==0.1.2
langchain-core==0.2.30
langchain-text-splitters==0.2.2
pymupdf==1.24.9
python-dotenv==1.0.1
streamlit==1.37.1
